Fix use of __builtin_assume_aligned for Clang 3.4
It was not yet supported in Clang 3.4, but fortunately there is a macro
to check whether it is supported. For now fall back to doing nothing,
like the code for non-gcc non-clang is doing already.

Additionally, it was the case that PIK_COMPILER_CLANG implied
PIK_COMPILER_GCC (because Clang pretends to be GCC for compatibility).
Change the check so Clang and GCC are now mutually exclusive. If this is
not desired, the check could be changed to check for Clang first.